TST: Modify conf.yaml for proper happi plugin syntax
I did not add a specific test for this as it looks like it was not done for the
other plugins. It might nice at some point to add tests that our conf.yaml is
converted to the expected dictionary structure. These kinds of mistake are
swallowed up by try/except blocks
the other plugins.