ENH: Helper function to find current log files
Instead of repeating the code to find a handler of a certain name, a generic
function was created to find a handler of any name. This is then used to find
the base log file name and search the directory for any other log messages that
may have been created.