ENH: Post all issues to Github using web.cfg info
Instead of saving the report as a JSON file then posting it with a separate
function, instead just post immediately after we gather the information. This
deprecates the `save_report` function. We also need to allow the opr accounts
to stash this information in the web.cfg file so that we don't have to
distribute a password.